requirement.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-genai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0 
scikit-learn>=1.3.0
rank-bm25>=0.2.2